fixed styles in not found page
