+++ client/src/App.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 
 import Login from './pages/Login';
-import NotFoundPage from './pages/404/404';
+import NotFoundPage from './pages/404';
 
 const App = () => {
   return (
