--- client/src/pages/404/index.css
@@ -1,26 +1,0 @@
-.notFoundPageTitle {
-    color: #737399;
-    font-size: 150px;
-    font-weight: bold;
-}
-
-.notFoundPageSubTitle {
-    font-size: 30px;
-    font-weight: bold;
-}
-
-.notFoundPageText{
-    font-size: 18px;
-}
-
-@media screen and (max-width: 450px) {
-    .notFoundPageTitle {
-        font-size: 90px;
-    }
-    .notFoundPageSubTitle {
-        font-size: 24px;
-    }
-    .notFoundPageText{
-        font-size: 16px;
-    }
-}
